very small optimizations VDP Text2 dirty colour table check
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@1193 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/SDLHiRenderer.hh
@@ -69,9 +69,9 @@ private:
 
 	inline void sync(const EmuTime &time);
 
-	inline void renderBitmapLines(int line, const EmuTime &until);
-	inline void renderPlanarBitmapLines(int line, const EmuTime &until);
-	inline void renderCharacterLines(int line, const EmuTime &until);
+	inline void renderBitmapLines(byte line, const EmuTime &until);
+	inline void renderPlanarBitmapLines(byte line, const EmuTime &until);
+	inline void renderCharacterLines(byte line, const EmuTime &until);
 
 	/** Get width of the left border in pixels.
 	  * This is equal to the X coordinate of the display area.
